# openapi/vocabularies.yaml
#
# Questo e' un progetto d'esempio in formato OpenAPI 3.0 API Starter Kit.
#
# Stai per sviluppare un API? Segui attentamente questo template e
#   potrai integrarla facilmente in Developers Italia.
#
# Se usi componenti in modo ricorrente, yaml
#   permette di definirle e riusarle.
#   Puoi metterle nell'oggetto x-commons e rimuoverlo
#   prima di pubblicarle.
x-commons:
  ratelimit-headers: &ratelimit-headers
    X-RateLimit-Limit:
      $ref: '#/components/headers/X-RateLimit-Limit'
    X-RateLimit-Remaining:
      $ref: '#/components/headers/X-RateLimit-Remaining'
    X-RateLimit-Reset:
      $ref: '#/components/headers/X-RateLimit-Reset'
  common-responses: &common-responses
    '400':
      $ref: '#/components/responses/400BadRequest'
    '429':
      $ref: '#/components/responses/429TooManyRequests'
    '503':
      $ref: '#/components/responses/503ServiceUnavailable'
    default:
      $ref: '#/components/responses/default'

openapi: 3.0.2
info:
  version: "0.0.1"
  title: |-
    Controlled Vocabularies.
  x-summary: >-
    Access controlled vocabularies via a REST interface.
  description: |
    #### Documentation
    This API exposes via a REST interface, dataset contained in RDF controlled vocabularies
    such as the list of foreign countries or the education levels.

    This API helps:

      * to uniform syntax and semantic of data used to create public services;
      * both public and private service providers in improving the interoperability of their services.

    Every API references the source of its data.

    #### Notes

    This API do not requires authentication or authorization,
    but must be used according the referenced terms of services, including
    the throttling and caching information.

    #### Examples

    Retrieve the list of exposed vocabularies via

    ```
    http https://localhost:8443/vocabularies/v1/vocabularies
    {
      'entries': [
        {"href": "https://"}
      ]
    }
    ```

  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: robipolli@gmail.com
    name: Roberto Polli
    url: https://twitter.com/ioggstream
  x-audience:
    - public
  # Ogni API deve avere un UUID, invariante nel tempo e
  #   rispetto al `title`.
  x-api-id: 00000000-0000-0000-0000-000000000000
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'

  # Per dare indicazioni sul ciclo di vita dell'API
  # si puo' usare un parametro come:
  x-lifecycle:
    published: 1970-01-01
    deprecated: 2050-01-01
    retired: 2050-06-01
    maturity: published

tags:
  - name: public
    description: Retrieve informations
    externalDocs:
      url: http://docs.my-api.com/pet-operations.htm
servers:
  - description: Development server
    url: /vocabularies/v1
  - description: Test server
    url: https://api.example.com/vocabularies/v1
    x-sandbox: true
    # Questo parametro indica come testare il servizio.
    x-healthCheck:
      url: https://api.example.com/vocabularies/v1/status
      interval: 300
      timeout: 15

paths:
  /vocabularies:
    get:
      summary: List vocabularies
      description: List vocabularies and their metadata.
      operationId: api.list_vocabularies
      tags:
      - public
      responses:
        <<: *common-responses
        '200':
          description: |
            Returned a list of vocabularies with their Metadata
            and the URL.

            TODO: A way of returing versioned resources must be identified.
          headers:
            <<: *ratelimit-headers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entries'

  /vocabularies/{vocabulary_id}/{entry_id}:
    get:
      summary: Return a given entry from a vocabulary
      description: |-
        A single entry from the selected vocabulary.
      operationId: api.get_entry
      tags:
      - public
      parameters:
        - name: vocabulary_id
          schema:
            type: string
            maxLength: 64
            pattern: >-
              [a-zA-Z0-9\-]{0,64}
          in: path
          required: true
        - name: entry_id
          schema:
            type: string
            maxLength: 64
          in: path
          required: true
        - name: format
          schema:
            type: string
            enum:
            - 'jsonld'
            - 'json'
          in: query
      responses:
        <<: *common-responses
        '200':
          description: |
            Il server ha ritornato il codepoint del vocabolario.
          headers:
            <<: *ratelimit-headers
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/Entry'
                - type: object
                  properties:
                    "@context":
                      type: object
                      maxProperties: 50
  /vocabularies/{vocabulary_id}:
    get:
      summary: Return one or more entries from a vocabulary.
      description: |
        Query data from the dataset using available fields.

        Invalid parameters result in a `400` response.
      operationId: api.list_entries
      tags:
        - public
      parameters:
        - name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            format: int32
          in: query
          required: false
        - name: cursor
          schema:
            type: string
            maxLength: 64
            minLength: 1
            pattern: >-
              [a-zA-Z0-9]{,64}
          in: query
          required: false
        - name: vocabulary_id
          schema:
            type: string
            maxLength: 64
          in: path
          required: true
        - name: label_it
          description: |-
            A string to be matched in the label_it of the vocabulary.
          allowReserved: false
          schema:
            type: string
            maxLength: 64
            pattern: >-
              %?[a-zA-Z0-9]{,62}%?
          in: query
          required: false
      responses:
        <<: *common-responses
        '200':
          description: |
            A list of the matching entries.
          headers:
            <<: *ratelimit-headers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entries'
  /status:
    get:
      summary: Ritorna lo stato dell'applicazione.
      description: |
        Ritorna lo stato dell'applicazione. A scopo
        di test, su base randomica puo' ritornare
        un errore.
      operationId: api.get_status
      tags:
        - public
      responses:
        <<: *common-responses
        '200':
          description: |
            Server returned correctly its status.
          headers:
            <<: *ratelimit-headers
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

components:

  headers:
    X-RateLimit-Limit:
      $ref: 'https://teamdigitale.github.io/openapi/0.0.7/definitions.yaml#/headers/X-RateLimit-Limit'
    X-RateLimit-Remaining:
      $ref: 'https://teamdigitale.github.io/openapi/0.0.7/definitions.yaml#/headers/X-RateLimit-Remaining'
    X-RateLimit-Reset:
      $ref: 'https://teamdigitale.github.io/openapi/0.0.7/definitions.yaml#/headers/X-RateLimit-Reset'
    Retry-After:
      $ref: 'https://teamdigitale.github.io/openapi/0.0.7/definitions.yaml#/headers/Retry-After'

  responses:
    # Predefined error codes for this API
    400BadRequest:
      $ref: 'https://teamdigitale.github.io/openapi/0.0.7/definitions.yaml#/responses/400BadRequest'
    404NotFound:
      $ref: 'https://teamdigitale.github.io/openapi/0.0.7/definitions.yaml#/responses/404NotFound'
    429TooManyRequests:
      $ref: 'https://teamdigitale.github.io/openapi/0.0.7/definitions.yaml#/responses/429TooManyRequests'
    503ServiceUnavailable:
      $ref: 'https://teamdigitale.github.io/openapi/0.0.7/definitions.yaml#/responses/503ServiceUnavailable'
    default:
      $ref: 'https://teamdigitale.github.io/openapi/0.0.7/definitions.yaml#/responses/default'

  schemas:
    Problem:
      $ref: 'https://teamdigitale.github.io/openapi/0.0.7/definitions.yaml#/schemas/Problem'
    Entry:
      type: object
      description: Un Timestamp in RFC5424
      additionalProperties: true
      properties:
        entry_id:
          type: string
          example: '42134fa'
    Entries:
      type: object
      description: a list of entries
      properties:
        entries:
          type: array
          items:
            $ref:  "#/components/schemas/Entry"
          maxItems: 100